conjugaciones_api/requirements.txt
asgiref==3.2.7
Django==3.0.5
django-cors-headers==3.2.1
django-redis==4.11.0
djangorestframework==3.11.0
djangorestframework-camel-case==1.1.2
psycopg2-binary==2.8.5
pytz==2019.3
redis==3.4.1
sqlparse==0.3.1
